requirements.txt
asgiref==3.8.1
asttokens==2.4.1
cfgv==3.4.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.6
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7
executing==2.0.1
filelock==3.14.0
identify==2.5.36
inflection==0.5.1
ipython==8.25.0
jedi==0.19.1
Markdown==3.6
matplotlib-inline==0.1.7
nodeenv==1.9.1
packaging==24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.46
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
uritemplate==4.1.1
virtualenv==20.26.2
wcwidth==0.2.13
